client: improve developer experience for failing search requests (#141)
Previously, for 403s we logged a generic error message about the domain
not being authorized, but that may not always be the case. Instead, now,
We log the error message from the server, plus the HTTP and gRPC code.
If there's a 403 error we also log another message telling the developer
they might need to check if the domain is an authorized query domain.

A test has been added which tests the client success/error states using
a real server (using msw) and using real fetch calls.

While writing a test I discovered `err instanceof RequestError` doesn't
work due to our compilation target so that has been fixed as well.